requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==1.0.1
Werkzeug==2.2.2
Pillow==9.2.0
python-dotenv==1.1.0
google-cloud-vision==3.3.0
openai==0.27.8
SQLAlchemy==1.4.49
WTForms==3.0.1